feat: Refactor editable parameters and routing table components for improved structure and functionality
