+++ src/graphics/renderables/routing_table.ts
